+++ src/mefit/Forms/frmEfiRom.Designer.cs
@@ -108,6 +108,7 @@ namespace Mac_EFI_Toolkit.Forms
             this.lblTitle = new Mac_EFI_Toolkit.UI.Controls.METLabel();
             this.pbxLogo = new System.Windows.Forms.PictureBox();
             this.tlpTitle = new System.Windows.Forms.TableLayoutPanel();
+            this.lblPagefile = new System.Windows.Forms.Label();
             this.cmdClose = new System.Windows.Forms.Button();
             this.cmdMinimize = new System.Windows.Forms.Button();
             this.helloToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
@@ -177,6 +178,7 @@ namespace Mac_EFI_Toolkit.Forms
             this.toolStripSeparator7 = new System.Windows.Forms.ToolStripSeparator();
             this.aboutToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
             this.settingsToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
+            this.manualToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
             this.pnlMain.SuspendLayout();
             this.tlpMain.SuspendLayout();
             this.tlpFilename.SuspendLayout();
@@ -1304,7 +1306,7 @@ namespace Mac_EFI_Toolkit.Forms
             this.lblTitle.Location = new System.Drawing.Point(44, 0);
             this.lblTitle.Margin = new System.Windows.Forms.Padding(0);
             this.lblTitle.Name = "lblTitle";
-            this.lblTitle.Size = new System.Drawing.Size(617, 50);
+            this.lblTitle.Size = new System.Drawing.Size(517, 50);
             this.lblTitle.TabIndex = 0;
             this.lblTitle.Text = "EFIROM";
             this.lblTitle.TextAlign = System.Drawing.ContentAlignment.MiddleLeft;
@@ -1327,15 +1329,17 @@ namespace Mac_EFI_Toolkit.Forms
             // 
             this.tlpTitle.BackColor = System.Drawing.Color.FromArgb(((int)(((byte)(15)))), ((int)(((byte)(15)))), ((int)(((byte)(15)))));
             this.tlpTitle.BackgroundImage = global::Mac_EFI_Toolkit.Properties.Resources.imgSprite;
-            this.tlpTitle.ColumnCount = 4;
+            this.tlpTitle.ColumnCount = 5;
             this.tlpTitle.ColumnStyles.Add(new System.Windows.Forms.ColumnStyle(System.Windows.Forms.SizeType.Absolute, 44F));
             this.tlpTitle.ColumnStyles.Add(new System.Windows.Forms.ColumnStyle(System.Windows.Forms.SizeType.Percent, 100F));
+            this.tlpTitle.ColumnStyles.Add(new System.Windows.Forms.ColumnStyle(System.Windows.Forms.SizeType.Absolute, 100F));
             this.tlpTitle.ColumnStyles.Add(new System.Windows.Forms.ColumnStyle(System.Windows.Forms.SizeType.Absolute, 50F));
             this.tlpTitle.ColumnStyles.Add(new System.Windows.Forms.ColumnStyle(System.Windows.Forms.SizeType.Absolute, 50F));
+            this.tlpTitle.Controls.Add(this.lblPagefile, 2, 0);
             this.tlpTitle.Controls.Add(this.lblTitle, 1, 0);
-            this.tlpTitle.Controls.Add(this.cmdClose, 3, 0);
+            this.tlpTitle.Controls.Add(this.cmdClose, 4, 0);
             this.tlpTitle.Controls.Add(this.pbxLogo, 0, 0);
-            this.tlpTitle.Controls.Add(this.cmdMinimize, 2, 0);
+            this.tlpTitle.Controls.Add(this.cmdMinimize, 3, 0);
             this.tlpTitle.Dock = System.Windows.Forms.DockStyle.Top;
             this.tlpTitle.Location = new System.Drawing.Point(1, 1);
             this.tlpTitle.Margin = new System.Windows.Forms.Padding(0);
@@ -1344,6 +1348,23 @@ namespace Mac_EFI_Toolkit.Forms
             this.tlpTitle.RowStyles.Add(new System.Windows.Forms.RowStyle(System.Windows.Forms.SizeType.Absolute, 50F));
             this.tlpTitle.Size = new System.Drawing.Size(761, 50);
             this.tlpTitle.TabIndex = 0;
+            // 
+            // lblPagefile
+            // 
+            this.lblPagefile.Anchor = System.Windows.Forms.AnchorStyles.Right;
+            this.lblPagefile.AutoSize = true;
+            this.lblPagefile.BackColor = System.Drawing.Color.FromArgb(((int)(((byte)(30)))), ((int)(((byte)(30)))), ((int)(((byte)(30)))));
+            this.lblPagefile.BorderStyle = System.Windows.Forms.BorderStyle.FixedSingle;
+            this.lblPagefile.Font = new System.Drawing.Font("Segoe UI Semibold", 9F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
+            this.lblPagefile.ForeColor = System.Drawing.Color.White;
+            this.lblPagefile.Location = new System.Drawing.Point(632, 14);
+            this.lblPagefile.Margin = new System.Windows.Forms.Padding(0, 0, 6, 0);
+            this.lblPagefile.Name = "lblPagefile";
+            this.lblPagefile.Size = new System.Drawing.Size(23, 22);
+            this.lblPagefile.TabIndex = 2;
+            this.lblPagefile.Text = "...";
+            this.lblPagefile.TextAlign = System.Drawing.ContentAlignment.MiddleCenter;
+            this.lblPagefile.Visible = false;
             // 
             // cmdClose
             // 
@@ -1950,13 +1971,14 @@ namespace Mac_EFI_Toolkit.Forms
             this.cmsHelp.ForeColor = System.Drawing.Color.FromArgb(((int)(((byte)(255)))), ((int)(((byte)(255)))), ((int)(((byte)(255)))));
             this.cmsHelp.ImageScalingSize = new System.Drawing.Size(20, 20);
             this.cmsHelp.Items.AddRange(new System.Windows.Forms.ToolStripItem[] {
+            this.manualToolStripMenuItem,
             this.viewApplicationLogToolStripMenuItem,
             this.toolStripSeparator7,
             this.aboutToolStripMenuItem,
             this.settingsToolStripMenuItem});
             this.cmsHelp.Name = "cmsHelp";
             this.cmsHelp.ShowImageMargin = false;
-            this.cmsHelp.Size = new System.Drawing.Size(252, 94);
+            this.cmsHelp.Size = new System.Drawing.Size(252, 150);
             // 
             // viewApplicationLogToolStripMenuItem
             // 
@@ -1989,6 +2011,14 @@ namespace Mac_EFI_Toolkit.Forms
             this.settingsToolStripMenuItem.Size = new System.Drawing.Size(251, 28);
             this.settingsToolStripMenuItem.Text = "Settings";
             this.settingsToolStripMenuItem.Click += new System.EventHandler(this.settingsToolStripMenuItem_Click);
+            // 
+            // manualToolStripMenuItem
+            // 
+            this.manualToolStripMenuItem.Name = "manualToolStripMenuItem";
+            this.manualToolStripMenuItem.ShortcutKeys = System.Windows.Forms.Keys.F1;
+            this.manualToolStripMenuItem.Size = new System.Drawing.Size(251, 28);
+            this.manualToolStripMenuItem.Text = "Manual";
+            this.manualToolStripMenuItem.Click += new System.EventHandler(this.manualToolStripMenuItem_Click);
             // 
             // frmEfiRom
             // 
@@ -2027,6 +2057,7 @@ namespace Mac_EFI_Toolkit.Forms
             this.tlpMenu.ResumeLayout(false);
             ((System.ComponentModel.ISupportInitialize)(this.pbxLogo)).EndInit();
             this.tlpTitle.ResumeLayout(false);
+            this.tlpTitle.PerformLayout();
             this.cmsCopy.ResumeLayout(false);
             this.cmsTools.ResumeLayout(false);
             this.cmsExport.ResumeLayout(false);
@@ -2174,5 +2205,7 @@ namespace Mac_EFI_Toolkit.Forms
         private System.Windows.Forms.ToolStripSeparator toolStripSeparator7;
         private System.Windows.Forms.ToolStripMenuItem aboutToolStripMenuItem;
         private System.Windows.Forms.ToolStripMenuItem settingsToolStripMenuItem;
+        private System.Windows.Forms.Label lblPagefile;
+        private System.Windows.Forms.ToolStripMenuItem manualToolStripMenuItem;
     }
 }